Fix build of types module with only json or only xml
Bump to 0.15.1
