+++ async-opcua-codegen/Cargo.toml
@@ -33,4 +33,4 @@ thiserror = "1.0.63"
 tracing = { workspace = true }
 uuid = "1.10.0"
 
-async-opcua-xml = { path = "../async-opcua-xml", version = "0.15.0" }
+async-opcua-xml = { path = "../async-opcua-xml", version = "0.15.1" }
