+++ async-opcua-nodes/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "async-opcua-nodes"
-version = "0.15.0"
+version = "0.15.1"
 description = "OPC UA node representation and import framework"
 authors = ["Adam Lock <locka99@gmail.com>", "Einar Omang <einar@omang.com>"]
 homepage = "https://github.com/freeopcua/async-opcua"
@@ -25,9 +25,9 @@ tracing = { workspace = true }
 regex = { workspace = true }
 thiserror = { workspace = true }
 
-async-opcua-macros = { path = "../async-opcua-macros", version = "0.15.0" }
-async-opcua-types = { path = "../async-opcua-types", version = "0.15.0" }
-async-opcua-xml = { path = "../async-opcua-xml", optional = true, version = "0.15.0" }
+async-opcua-macros = { path = "../async-opcua-macros", version = "0.15.1" }
+async-opcua-types = { path = "../async-opcua-types", version = "0.15.1" }
+async-opcua-xml = { path = "../async-opcua-xml", optional = true, version = "0.15.1" }
 
 [dev-dependencies]
 async-opcua-nodes = { path = ".", features = ["xml"] }
