+++ async-opcua-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "async-opcua-types"
-version = "0.15.0"
+version = "0.15.1"
 description = "OPC UA data types"
 authors = ["Adam Lock <locka99@gmail.com>", "Einar Omang <einar@omang.com>"]
 homepage = "https://github.com/freeopcua/async-opcua"
@@ -32,8 +32,8 @@ thiserror = { workspace = true }
 tracing = { workspace = true }
 uuid = { workspace = true }
 
-async-opcua-macros = { path = "../async-opcua-macros", version = "0.15.0" }
-async-opcua-xml = { path = "../async-opcua-xml", optional = true, version = "0.15.0" }
+async-opcua-macros = { path = "../async-opcua-macros", version = "0.15.1" }
+async-opcua-xml = { path = "../async-opcua-xml", optional = true, version = "0.15.1" }
 
 [dev-dependencies]
 async-opcua-types = { path = ".", features = ["xml", "json"] }
