+++ samples/custom-structures-client/Cargo.toml
@@ -12,7 +12,7 @@ env_logger = { workspace = true }
 
 [dependencies.async-opcua]
 path = "../../async-opcua"
-version = "0.15.0"
+version = "0.15.1"
 features = ["client"]
 default-features = false
